requirements.txt
absl-py==0.1.13
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.8
anaconda-project==0.8.0
aniso8601==2.0.1
asn1crypto==0.22.0
astor==0.6.2
astroid==1.5.3
astropy==2.0.2
babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.0.0
bokeh==0.12.7
boto==2.48.0
Bottleneck==1.2.1
CacheControl==0.12.3
cairocffi==0.8.0
CairoSVG==2.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.2
conda==4.3.27
conda-build==3.0.22
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.0.3
cssselect2==0.2.1
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.2
datashape==0.5.4
decorator==4.1.2
defusedxml==0.5.0
distlib==0.2.5
distributed==1.18.3
Django==2.0
django-ckeditor==5.4.0
django-jet==1.0.7
django-js-asset==1.0.0
django-pageviews==1.0.0
django-taggit==0.22.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
Fabric==1.14.0
Faker==0.8.5
fastcache==1.0.2
fbchat==1.3.1
feedparser==5.2.1
filelock==2.0.12
Flask==0.12.2
Flask-Cors==3.0.3
Flask-PyMongo==0.5.1
Flask-RESTful==0.3.6
gast==0.2.0
gevent==1.2.2
glob2==0.5
google-api-python-client==1.6.5
greenlet==0.4.12
grpcio==1.10.0
h5py==2.7.0
heapdict==1.0.0
html5lib==0.999999999
httplib2==0.10.3
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
lazy-object-proxy==1.3.1
llvmlite==0.20.0
lml==0.0.1
locket==0.2.0
lockfile==0.12.2
lxml==3.8.0
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
menuinst==1.4.8
mistune==0.7.4
mongoengine==0.14.3
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0+10.g143f70e
numexpr==2.6.2
numpy==1.13.3+mkl
numpydoc==0.7.0
oauth2client==4.1.2
oauthlib==2.0.6
odo==0.5.1
olefile==0.44
opencv-python==3.3.0
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
paramiko==2.4.0
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pdfcrowd==4.3.0
pdfminer3k==1.3.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
pkginfo==1.4.1
ply==3.10
progress==1.3
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==5.2.2
py==1.4.34
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycosat==0.6.2
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pyexcel==0.5.7
pyexcel-io==0.5.6
pyexcel-xlsx==0.5.5
pyflakes==1.5.0
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.7.2
pymongo==3.5.1
PyNaCl==1.2.1
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
Pyphen==0.9.4
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
PyWavelets==0.5.2
pywin32==221
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
requests-oauthlib==0.8.0
rope==0.10.5
rsa==3.4.2
ruamel-yaml==0.11.14
schedule==0.5.0
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.1
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.3
SQLAlchemy==1.1.13
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
termcolor==1.1.0
testpath==0.3.1
texttable==1.2.1
tinycss2==0.6.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
vboxapi==1.0
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wkhtmltopdf==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==0.9.8
xlwings==0.11.4
xlwt==1.3.0
youtube-dl==2017.12.31
zict==0.1.2
